requirements.txt
pytz==2022.7.1
numpy==1.25.0
pandas==1.5.3
psycopg2==2.9.3
yfinance==0.2.12
requests==2.28.1
sqlalchemy==1.4.39
progressbar2==4.2.0